career-assistant/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.6.1
langchain==0.0.340
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.4
python-docx==1.1.0
numpy==1.24.4
scikit-learn==1.3.2
pandas==2.1.4
requests==2.31.0
celery==5.3.4
redis==5.0.1
schedule==1.2.0
beautifulsoup4==4.12.2
selenium==4.15.2
tqdm==4.66.1